requirements.txt
weasyprint==68.1
python-docx==1.2.0
